Graphviz generator
Generates a graphviz of the tree corresponding to the statemachine and
its underlying states.

This is a hierarchy-graph, not a finite state machine. It does not
display the transitions.  It is however more compact, and hopefully
is easier to follow on-line.

Displaying of the current active state is not yet implemented.